examples/strands_officebench_agent/pyproject.toml
[project]
name = "strands-officebench-agent-example"
version = "0.1.0"
description = "Example: Strands OfficeBench Agent using Bedrock AgentCore RL Toolkit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bedrock-agentcore>=1.0.3",
    "bedrock-agentcore-starter-toolkit>=0.2.0",
    "boto3>=1.40.55",
    "python-dotenv>=1.0.0",
    "strands-agents[openai]>=1.18.0",
    "strands-agents-tools>=0.2.16",
    # OfficeBench evaluation deps
    "openpyxl",
    "python-docx",
    "PyMuPDF",
    "pdf2docx",
    "pytesseract",
    "icalendar",
    "Pillow",
    "pytz",
    "fire",
]

[tool.setuptools]
py-modules = ["dev_app", "reward", "models", "utils", "tools", "deploy"]
